Inline payment and user operation as before
